Move sample config to separate file
