+++ plugins/processors/date/date_sample_config.go
@@ -1,0 +1,6 @@
+// DON'T EDIT; This file is used as a template by tools/generate_plugindata
+package date
+
+func (d *Date) SampleConfig() string {
+	return `{{ .SampleConfig }}`
+}
