+++ plugins/processors/override/override_sample_config.go
@@ -1,0 +1,6 @@
+// DON'T EDIT; This file is used as a template by tools/generate_plugindata
+package override
+
+func (p *Override) SampleConfig() string {
+	return `{{ .SampleConfig }}`
+}
