+++ plugins/processors/parser/parser_sample_config.go
@@ -1,0 +1,6 @@
+// DON'T EDIT; This file is used as a template by tools/generate_plugindata
+package parser
+
+func (p *Parser) SampleConfig() string {
+	return `{{ .SampleConfig }}`
+}
