+++ plugins/processors/port_name/port_name_sample_config.go
@@ -1,0 +1,6 @@
+// DON'T EDIT; This file is used as a template by tools/generate_plugindata
+package port_name
+
+func (pn *PortName) SampleConfig() string {
+	return `{{ .SampleConfig }}`
+}
