+++ plugins/processors/reverse_dns/reverse_dns_sample_config.go
@@ -1,0 +1,6 @@
+// DON'T EDIT; This file is used as a template by tools/generate_plugindata
+package reverse_dns
+
+func (r *ReverseDNS) SampleConfig() string {
+	return `{{ .SampleConfig }}`
+}
