+++ plugins/processors/s2geo/s2geo_sample_config.go
@@ -1,0 +1,6 @@
+// DON'T EDIT; This file is used as a template by tools/generate_plugindata
+package s2geo
+
+func (g *Geo) SampleConfig() string {
+	return `{{ .SampleConfig }}`
+}
