+++ plugins/processors/tag_limit/tag_limit_sample_config.go
@@ -1,0 +1,6 @@
+// DON'T EDIT; This file is used as a template by tools/generate_plugindata
+package tag_limit
+
+func (d *TagLimit) SampleConfig() string {
+	return `{{ .SampleConfig }}`
+}
