+++ plugins/processors/unpivot/unpivot_sample_config.go
@@ -1,0 +1,6 @@
+// DON'T EDIT; This file is used as a template by tools/generate_plugindata
+package unpivot
+
+func (p *Unpivot) SampleConfig() string {
+	return `{{ .SampleConfig }}`
+}
